done 04.02.2020

+++ db/CREATE_TABLE_product.sql
@@ -2,6 +2,6 @@ CREATE TABLE product (
   product_id SERIAL PRIMARY KEY NOT NULL,
   name varchar(40) NOT NULL,
   description varchar(80) NOT NULL,
-  price integer NOT NULL,
+  price float NOT NULL,
   image_url text NOT NULL
 );
